d/rules: makefile definitions can't be recursive - sys_systems for s390x
We had this as a direct change in the past.
The Vendor dependent change got added on pushing it to Debian and tested
for Debian on that. Now that we build Ubuntu with it it fails with:
 debian/rules:73: *** Recursive variable 'sys_systems' references itself

Easy fix that will eventually go to Debian as well to drop the Delta but
isn't needed urgently there as they are not entering the if-clause.

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>